Added webpack configuration.
