+++ package.json
@@ -1,0 +1,28 @@
+{
+  "name": "Scatter-Brain",
+  "version": "0.0.1",
+  "description": "Easy Way to tag your thoughts.",
+  "author": "Abhimanyu Babbar <babbar@kth.se>",
+  "dependencies": {
+    "react": "^0.14.6",
+    "react-modal": "latest",
+    "underscore": "latest"
+  },
+  "main": "index.js",
+  "devDependencies": {
+    "jsx-loader": "^0.13.2",
+    "webpack": "^1.12.12",
+    "webpack-dev-server": "^1.14.1"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/babbarshaer/scatter-brain.git"
+  },
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/babbarshaer/scatter-brain/issues"
+  }
+}
