+++ public/index.html
@@ -18,7 +18,7 @@
     <div id="container" class="container"></div>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/react/0.14.3/react.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/react/0.14.3/react-dom.js"></script>
-    <script type="text/babel" src="/js/script.js"></script>
+    <script src="/js/bundle.js"></script>
 </body>
 
 </html>
