+++ webpack.config.js
@@ -1,0 +1,26 @@
+module.exports = {
+    entry: './public/js/index.jsx',
+    output: {
+        filename: './public/js/bundle.js', //this is the default name, so you can skip it
+            //at this directory our bundle file will be available
+            //make sure port 8090 is used when launching webpack-dev-server
+            publicPath: 'http://localhost:8090/assets'
+        },
+    module: {
+            loaders: [
+                        {
+                                        //tell webpack to use jsx-loader for all *.jsx files
+                                        test: /\.jsx$/,
+                                        loader: 'jsx-loader?insertPragma=React.DOM&harmony'
+                                    }
+                    ]
+        },
+    externals: {
+            //don't bundle the 'react' npm package with our bundle.js
+            //but get it from a global 'React' variable
+            'react': 'React'
+        },
+    resolve: {
+            extensions: ['', '.js', '.jsx']
+        }
+};
